Update README.md
jQuery SMCstylus Circle Timers is a jQuery plugin that provides beautiful animated countdowns in circle shapes.

Features: You can fill the center of circles, the past time lines and the left time lines with solid or gradient colors. There is also the possibility to change the cap style of animated line. Add separate shadow (color and diffuse) for main timer and left time lines and more...